backend/requirements.txt
fastapi[standard]
uvicorn
ultralytics
opencv-python-headless
groq
python-dotenv
websockets
